Move SingleStreamConnectionTask and MultiStreamConnectionTask to submodules
