+++ src/libp2p/collection.rs
@@ -67,6 +67,12 @@ pub use super::peer_id::PeerId;
 pub use super::read_write::ReadWrite;
 pub use established::{ConfigRequestResponse, ConfigRequestResponseIn, InboundError};
 
+pub use multi_stream::MultiStreamConnectionTask;
+pub use single_stream::SingleStreamConnectionTask;
+
+mod multi_stream;
+mod single_stream;
+
 /// Configuration for a [`Network`].
 pub struct Config {
     /// Seed for the randomness within the networking state machine.
@@ -327,21 +333,15 @@ where
         let connection_id = self.next_connection_id;
         self.next_connection_id.0 += 1;
 
-        let connection_task = SingleStreamConnectionTask {
-            connection: SingleStreamConnectionTaskInner::Handshake {
-                handshake: handshake::HealthyHandshake::new(is_initiator),
-                randomness_seed: self.randomness_seeds.gen(),
-                timeout: when_connected + self.handshake_timeout,
-                noise_key: self.noise_key.clone(),
-                notification_protocols: self.notification_protocols.clone(),
-                request_response_protocols: self.request_response_protocols.clone(),
-                ping_protocol: self.ping_protocol.clone(),
-            },
-            pending_messages: VecDeque::with_capacity({
-                // We never buffer more than a few messages.
-                4
-            }),
-        };
+        let connection_task = SingleStreamConnectionTask::new(
+            self.randomness_seeds.gen(),
+            is_initiator,
+            when_connected + self.handshake_timeout,
+            self.noise_key.clone(),
+            self.notification_protocols.clone(),
+            self.request_response_protocols.clone(),
+            self.ping_protocol.clone(),
+        );
 
         let _previous_value = self.connections.insert(
             connection_id,
@@ -369,43 +369,13 @@ where
         let connection_id = self.next_connection_id;
         self.next_connection_id.0 += 1;
 
-        let connection_task = MultiStreamConnectionTask {
-            connection: MultiStreamConnectionTaskInner::Established {
-                established: established::MultiStream::new(established::Config {
-                    notifications_protocols: self
-                        .notification_protocols
-                        .iter()
-                        .flat_map(|net| {
-                            let max_handshake_size = net.config.max_handshake_size;
-                            let max_notification_size = net.config.max_notification_size;
-                            iter::once(&net.config.protocol_name)
-                                .chain(net.config.fallback_protocol_names.iter())
-                                .map(move |name| {
-                                    established::ConfigNotifications {
-                                        name: name.clone(), // TODO: cloning :-/
-                                        max_handshake_size,
-                                        max_notification_size,
-                                    }
-                                })
-                        })
-                        .collect(),
-                    request_protocols: self.request_response_protocols.to_vec(), // TODO: overhead
-                    randomness_seed: self.randomness_seeds.gen(),
-                    ping_protocol: self.ping_protocol.to_string(), // TODO: cloning :-/
-                    ping_interval: Duration::from_secs(20),        // TODO: hardcoded
-                    ping_timeout: Duration::from_secs(10),         // TODO: hardcoded
-                    first_out_ping: now + Duration::from_secs(2),  // TODO: hardcoded
-                }),
-                outbound_substreams_map: hashbrown::HashMap::with_capacity_and_hasher(
-                    0,
-                    Default::default(),
-                ), // TODO: capacity?
-                outbound_substreams_reverse: hashbrown::HashMap::with_capacity_and_hasher(
-                    0,
-                    Default::default(),
-                ), // TODO: capacity?
-            },
-        };
+        let connection_task = MultiStreamConnectionTask::new(
+            self.randomness_seeds.gen(),
+            now,
+            self.notification_protocols.clone(),
+            self.request_response_protocols.clone(),
+            self.ping_protocol.clone(),
+        );
 
         let _previous_value = self.connections.insert(
             connection_id,
@@ -1314,1202 +1284,6 @@ pub struct ConnectionState {
 
     /// If `true`, the connection is shutting down.
     pub shutting_down: bool,
-}
-
-/// State machine dedicated to a single single-stream connection.
-pub struct SingleStreamConnectionTask<TNow> {
-    /// State machine of the underlying connection.
-    connection: SingleStreamConnectionTaskInner<TNow>,
-
-    /// Buffer of messages destined to the coordinator.
-    ///
-    /// Never goes above a few elements.
-    pending_messages: VecDeque<ConnectionToCoordinatorInner>,
-}
-
-enum SingleStreamConnectionTaskInner<TNow> {
-    /// Connection is still in its handshake phase.
-    Handshake {
-        handshake: handshake::HealthyHandshake,
-
-        /// Seed that will be used to initialize randomness when building the
-        /// [`established::SingleStream`].
-        /// This seed is computed during the handshake in order to avoid having to access a shared
-        /// state when the handshake is over. While it seems a bit dangerous to leave a randomness
-        /// seed in plain memory, the randomness isn't used for anything critical or related to
-        /// cryptography, but only for example to avoid hash collision attacks.
-        randomness_seed: [u8; 32],
-
-        /// When the handshake phase times out.
-        timeout: TNow,
-
-        /// See [`Config::noise_key`].
-        noise_key: Arc<NoiseKey>,
-
-        /// See [`OverlayNetwork`].
-        notification_protocols: Arc<[OverlayNetwork]>,
-
-        /// See [`Config::request_response_protocols`].
-        request_response_protocols: Arc<[ConfigRequestResponse]>,
-
-        /// See [`Config::ping_protocol`].
-        ping_protocol: Arc<str>,
-    },
-
-    /// Connection has been fully established.
-    Established {
-        // TODO: user data of request redundant with the substreams mapping below
-        established: established::SingleStream<TNow, SubstreamId, ()>,
-
-        /// Because outgoing substream ids are assigned by the coordinator, we maintain a mapping
-        /// of the "outer ids" to "inner ids".
-        outbound_substreams_map:
-            hashbrown::HashMap<SubstreamId, established::SubstreamId, fnv::FnvBuildHasher>,
-
-        /// Reverse mapping.
-        // TODO: could be user datas in established?
-        outbound_substreams_reverse:
-            hashbrown::HashMap<established::SubstreamId, SubstreamId, fnv::FnvBuildHasher>,
-    },
-
-    /// Connection has finished its shutdown. A [`ConnectionToCoordinatorInner::ShutdownFinished`]
-    /// message has been sent and is waiting to be acknowledged.
-    ShutdownWaitingAck {
-        /// If true, [`SingleStreamConnectionTask::reset`] has been called. This doesn't modify
-        /// any of the behavior but is used to make sure that the API is used correctly.
-        was_api_reset: bool,
-    },
-
-    /// Connection has finished its shutdown and its shutdown has been acknowledged. There is
-    /// nothing more to do except stop the connection task.
-    ShutdownAcked {
-        /// If true, [`SingleStreamConnectionTask::reset`] has been called. This doesn't modify
-        /// any of the behavior but is used to make sure that the API is used correctly.
-        was_api_reset: bool,
-    },
-
-    /// Temporary state used to satisfy the borrow checker during state transitions.
-    Poisoned,
-}
-
-impl<TNow> SingleStreamConnectionTask<TNow>
-where
-    TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,
-{
-    /// Pulls a message to send back to the coordinator.
-    ///
-    /// This function takes ownership of `self` and optionally yields it back. If the first
-    /// option contains `None`, then no more message will be generated and the
-    /// [`SingleStreamConnectionTask`] has vanished. This will happen after the connection has been
-    /// shut down or reset.
-    /// It is possible for `self` to not be yielded back even if the [`ReadWrite`] that was last
-    /// passed to [`SingleStreamConnectionTask::read_write`] is still fully open, in which case the
-    /// API user should abruptly reset the connection, for example by sending a TCP RST flag. This
-    /// can happen for example if the connection seems unresponsive and that an attempt at closing
-    /// the connection in a clean way is futile.
-    ///
-    /// If any message is returned, it is the responsibility of the API user to send it to the
-    /// coordinator by calling [`Network::inject_connection_message`].
-    /// Do not attempt to buffer the message being returned, as it would work against the
-    /// back-pressure strategy used internally. As soon as a message is returned, it should be
-    /// delivered. If the coordinator is busy at the moment a message should be delivered, then
-    /// the entire thread of execution dedicated to this [`SingleStreamConnectionTask`] should be
-    /// paused until the coordinator is ready and the message delivered.
-    ///
-    /// Messages aren't generated spontaneously. In other words, you don't need to periodically
-    /// call this function just in case there's a new message. Messages are always generated after
-    /// either [`SingleStreamConnectionTask::read_write`] or [`SingleStreamConnectionTask::reset`]
-    /// has been called. Multiple messages can happen in a row.
-    ///
-    /// Because this function frees space in a buffer, calling
-    /// [`SingleStreamConnectionTask::read_write`] again after it has returned might read/write
-    /// more data and generate an event again. In other words, the API user should call
-    /// [`SingleStreamConnectionTask::read_write`] and
-    /// [`SingleStreamConnectionTask::pull_message_to_coordinator`] repeatedly in a loop until no
-    /// more message is generated.
-    pub fn pull_message_to_coordinator(
-        mut self,
-    ) -> (Option<Self>, Option<ConnectionToCoordinator>) {
-        // To be sure that there is no bug in the implementation, we make sure that the number of
-        // buffered messages doesn't go above a certain small limit.
-        debug_assert!(self.pending_messages.len() < 8);
-
-        let message = self
-            .pending_messages
-            .pop_front()
-            .map(|inner| ConnectionToCoordinator { inner });
-
-        // The `ShutdownAcked` state causes the task to exit.
-        let self_ret = if !matches!(
-            self.connection,
-            SingleStreamConnectionTaskInner::ShutdownAcked { .. }
-        ) {
-            Some(self)
-        } else {
-            None
-        };
-
-        (self_ret, message)
-    }
-
-    /// Injects a message that has been pulled using [`Network::pull_message_to_connection`].
-    ///
-    /// Calling this function might generate data to send to the connection. You should call
-    /// [`SingleStreamConnectionTask::read_write`] after this function has returned (unless you
-    /// have called [`SingleStreamConnectionTask::reset`] in the past).
-    pub fn inject_coordinator_message(&mut self, message: CoordinatorToConnection<TNow>) {
-        match (message.inner, &mut self.connection) {
-            (
-                CoordinatorToConnectionInner::StartRequest {
-                    request_data,
-                    timeout,
-                    protocol_index,
-                    substream_id,
-                },
-                SingleStreamConnectionTaskInner::Established {
-                    established,
-                    outbound_substreams_map,
-                    outbound_substreams_reverse,
-                },
-            ) => {
-                let inner_substream_id =
-                    established.add_request(protocol_index, request_data, timeout, substream_id);
-                let _prev_value = outbound_substreams_map.insert(substream_id, inner_substream_id);
-                debug_assert!(_prev_value.is_none());
-                let _prev_value =
-                    outbound_substreams_reverse.insert(inner_substream_id, substream_id);
-                debug_assert!(_prev_value.is_none());
-            }
-            (
-                CoordinatorToConnectionInner::OpenOutNotifications {
-                    handshake,
-                    now,
-                    overlay_network_index,
-                    substream_id: outer_substream_id,
-                },
-                SingleStreamConnectionTaskInner::Established {
-                    established,
-                    outbound_substreams_map,
-                    outbound_substreams_reverse,
-                },
-            ) => {
-                let inner_substream_id = established.open_notifications_substream(
-                    now,
-                    overlay_network_index,
-                    handshake,
-                    (),
-                );
-
-                let _prev_value =
-                    outbound_substreams_map.insert(outer_substream_id, inner_substream_id);
-                debug_assert!(_prev_value.is_none());
-                let _prev_value =
-                    outbound_substreams_reverse.insert(inner_substream_id, outer_substream_id);
-                debug_assert!(_prev_value.is_none());
-            }
-            (
-                CoordinatorToConnectionInner::CloseOutNotifications { substream_id },
-                SingleStreamConnectionTaskInner::Established {
-                    established,
-                    outbound_substreams_map,
-                    outbound_substreams_reverse,
-                },
-            ) => {
-                // It is possible that the remote has closed the outbound notification substream
-                // while the `CloseOutNotifications` message was being delivered, or that the API
-                // user close the substream before the message about the substream being closed
-                // was delivered to the coordinator.
-                if let Some(inner_substream_id) = outbound_substreams_map.remove(&substream_id) {
-                    outbound_substreams_reverse.remove(&inner_substream_id);
-                    established.close_notifications_substream(inner_substream_id);
-                }
-            }
-            (
-                CoordinatorToConnectionInner::QueueNotification {
-                    substream_id,
-                    notification,
-                },
-                SingleStreamConnectionTaskInner::Established {
-                    established,
-                    outbound_substreams_map,
-                    ..
-                },
-            ) => {
-                // It is possible that the remote has closed the outbound notification substream
-                // while a `QueueNotification` message was being delivered, or that the API user
-                // queued a notification before the message about the substream being closed was
-                // delivered to the coordinator.
-                // If that happens, we intentionally silently discard the message, causing the
-                // notification to not be sent. This is consistent with the guarantees about
-                // notifications delivered that are documented in the public API.
-                if let Some(inner_substream_id) = outbound_substreams_map.get(&substream_id) {
-                    established.write_notification_unbounded(*inner_substream_id, notification);
-                }
-            }
-            (
-                CoordinatorToConnectionInner::AnswerRequest {
-                    substream_id,
-                    response,
-                },
-                SingleStreamConnectionTaskInner::Established { established, .. },
-            ) => match established.respond_in_request(substream_id, response) {
-                Ok(()) => {}
-                Err(established::RespondInRequestError::SubstreamClosed) => {
-                    // As documented, answering an obsolete request is simply ignored.
-                }
-            },
-            (
-                CoordinatorToConnectionInner::AcceptInNotifications {
-                    substream_id,
-                    handshake,
-                },
-                SingleStreamConnectionTaskInner::Established { established, .. },
-            ) => {
-                // TODO: must verify that the substream is still valid
-                established.accept_in_notifications_substream(substream_id, handshake, ());
-            }
-            (
-                CoordinatorToConnectionInner::RejectInNotifications { substream_id },
-                SingleStreamConnectionTaskInner::Established { established, .. },
-            ) => {
-                // TODO: must verify that the substream is still valid
-                established.reject_in_notifications_substream(substream_id);
-            }
-            (
-                CoordinatorToConnectionInner::StartShutdown { .. },
-                SingleStreamConnectionTaskInner::Established { .. }
-                | SingleStreamConnectionTaskInner::Handshake { .. },
-            ) => {
-                // TODO: implement proper shutdown
-                self.pending_messages
-                    .push_back(ConnectionToCoordinatorInner::StartShutdown);
-                self.pending_messages
-                    .push_back(ConnectionToCoordinatorInner::ShutdownFinished);
-                self.connection = SingleStreamConnectionTaskInner::ShutdownWaitingAck {
-                    was_api_reset: false,
-                };
-            }
-            (
-                CoordinatorToConnectionInner::AcceptInNotifications { .. }
-                | CoordinatorToConnectionInner::RejectInNotifications { .. }
-                | CoordinatorToConnectionInner::AnswerRequest { .. }
-                | CoordinatorToConnectionInner::OpenOutNotifications { .. }
-                | CoordinatorToConnectionInner::CloseOutNotifications { .. }
-                | CoordinatorToConnectionInner::QueueNotification { .. },
-                SingleStreamConnectionTaskInner::Handshake { .. }
-                | SingleStreamConnectionTaskInner::ShutdownAcked { .. },
-            ) => unreachable!(),
-            (
-                CoordinatorToConnectionInner::AcceptInNotifications { .. }
-                | CoordinatorToConnectionInner::RejectInNotifications { .. }
-                | CoordinatorToConnectionInner::AnswerRequest { .. }
-                | CoordinatorToConnectionInner::OpenOutNotifications { .. }
-                | CoordinatorToConnectionInner::CloseOutNotifications { .. }
-                | CoordinatorToConnectionInner::QueueNotification { .. },
-                SingleStreamConnectionTaskInner::ShutdownWaitingAck { .. },
-            ) => {
-                // There might still be some messages coming from the coordinator after the
-                // connection task has sent a message indicating that it has shut down. This is
-                // due to the concurrent nature of the API and doesn't indicate a bug. These
-                // messages are simply ignored by the connection task.
-            }
-            (
-                CoordinatorToConnectionInner::ShutdownFinishedAck,
-                SingleStreamConnectionTaskInner::ShutdownWaitingAck {
-                    was_api_reset: was_reset,
-                },
-            ) => {
-                self.connection = SingleStreamConnectionTaskInner::ShutdownAcked {
-                    was_api_reset: *was_reset,
-                };
-            }
-            _ => todo!(), // TODO:
-        }
-    }
-
-    /// Sets the state of the connection to "reset".
-    ///
-    /// This should be called if the remote abruptly closes the connection, such as with a TCP/IP
-    /// RST flag.
-    ///
-    /// After this function has been called, it is illegal to call
-    /// [`SingleStreamConnectionTask::read_write`] or [`SingleStreamConnectionTask::reset`] again.
-    ///
-    /// Calling this function might have generated messages for the coordinator.
-    /// [`SingleStreamConnectionTask::pull_message_to_coordinator`] should be called afterwards in
-    /// order to process these messages.
-    ///
-    /// # Panic
-    ///
-    /// Panics if [`SingleStreamConnectionTask::reset`] has been called in the past.
-    ///
-    pub fn reset(&mut self) {
-        // It is illegal to call `reset` a second time. Verify that the user didn't do this.
-        if let SingleStreamConnectionTaskInner::ShutdownWaitingAck {
-            was_api_reset: true,
-        }
-        | SingleStreamConnectionTaskInner::ShutdownAcked {
-            was_api_reset: true,
-        } = self.connection
-        {
-            panic!()
-        }
-
-        self.pending_messages
-            .push_back(ConnectionToCoordinatorInner::StartShutdown);
-        self.pending_messages
-            .push_back(ConnectionToCoordinatorInner::ShutdownFinished);
-        self.connection = SingleStreamConnectionTaskInner::ShutdownWaitingAck {
-            was_api_reset: true,
-        };
-    }
-
-    /// Reads data coming from the connection, updates the internal state machine, and writes data
-    /// destined to the connection through the [`ReadWrite`].
-    ///
-    /// Calling this function might have generated messages for the coordinator.
-    /// [`SingleStreamConnectionTask::pull_message_to_coordinator`] should be called afterwards in
-    /// order to process these messages.
-    ///
-    /// # Panic
-    ///
-    /// Panics if [`SingleStreamConnectionTask::reset`] has been called in the past.
-    ///
-    pub fn read_write(&mut self, read_write: &'_ mut ReadWrite<'_, TNow>) {
-        // There is already at least one pending message. We back-pressure the connection by not
-        // performing any reading or writing, as this might generate more messages and open the
-        // door for a DoS attack by the remote. As documented, the API user is supposed to pull
-        // messages after this function has returned, meaning that they will drain the messages.
-        if !self.pending_messages.is_empty() {
-            return;
-        }
-
-        match mem::replace(
-            &mut self.connection,
-            SingleStreamConnectionTaskInner::Poisoned,
-        ) {
-            SingleStreamConnectionTaskInner::Established {
-                established,
-                mut outbound_substreams_map,
-                mut outbound_substreams_reverse,
-            } => match established.read_write(read_write) {
-                Ok((connection, event)) => {
-                    if read_write.is_dead() && event.is_none() {
-                        // TODO: provide error
-                        self.pending_messages
-                            .push_back(ConnectionToCoordinatorInner::StartShutdown);
-                        self.pending_messages
-                            .push_back(ConnectionToCoordinatorInner::ShutdownFinished);
-                        self.connection = SingleStreamConnectionTaskInner::ShutdownWaitingAck {
-                            was_api_reset: false,
-                        };
-                        return;
-                    }
-
-                    match event {
-                        Some(established::Event::InboundError(err)) => {
-                            self.pending_messages
-                                .push_back(ConnectionToCoordinatorInner::InboundError(err));
-                        }
-                        Some(established::Event::RequestIn {
-                            id,
-                            protocol_index,
-                            request,
-                        }) => {
-                            self.pending_messages.push_back(
-                                ConnectionToCoordinatorInner::RequestIn {
-                                    id,
-                                    protocol_index,
-                                    request,
-                                },
-                            );
-                        }
-                        Some(established::Event::Response { id, response, .. }) => {
-                            let outer_substream_id =
-                                outbound_substreams_reverse.remove(&id).unwrap();
-                            outbound_substreams_map.remove(&outer_substream_id).unwrap();
-                            self.pending_messages.push_back(
-                                ConnectionToCoordinatorInner::Response {
-                                    response,
-                                    id: outer_substream_id,
-                                },
-                            );
-                        }
-                        Some(established::Event::NotificationsInOpen {
-                            id,
-                            protocol_index,
-                            handshake,
-                        }) => {
-                            self.pending_messages.push_back(
-                                ConnectionToCoordinatorInner::NotificationsInOpen {
-                                    id,
-                                    protocol_index,
-                                    handshake,
-                                },
-                            );
-                        }
-                        Some(established::Event::NotificationsInOpenCancel { id, .. }) => {
-                            self.pending_messages.push_back(
-                                ConnectionToCoordinatorInner::NotificationsInOpenCancel { id },
-                            );
-                        }
-                        Some(established::Event::NotificationIn { id, notification }) => {
-                            self.pending_messages.push_back(
-                                ConnectionToCoordinatorInner::NotificationIn { id, notification },
-                            );
-                        }
-                        Some(established::Event::NotificationsInClose { id, outcome, .. }) => {
-                            self.pending_messages.push_back(
-                                ConnectionToCoordinatorInner::NotificationsInClose { id, outcome },
-                            );
-                        }
-                        Some(established::Event::NotificationsOutResult { id, result }) => {
-                            let outer_substream_id = *outbound_substreams_reverse.get(&id).unwrap();
-
-                            if result.is_err() {
-                                outbound_substreams_map.remove(&outer_substream_id);
-                                outbound_substreams_reverse.remove(&id);
-                            }
-
-                            self.pending_messages.push_back(
-                                ConnectionToCoordinatorInner::NotificationsOutResult {
-                                    id: outer_substream_id,
-                                    result: result
-                                        .map_err(|(err, _)| NotificationsOutErr::Substream(err)),
-                                },
-                            );
-                        }
-                        Some(established::Event::NotificationsOutCloseDemanded { id }) => {
-                            let outer_substream_id = *outbound_substreams_reverse.get(&id).unwrap();
-                            self.pending_messages.push_back(
-                                ConnectionToCoordinatorInner::NotificationsOutCloseDemanded {
-                                    id: outer_substream_id,
-                                },
-                            );
-                        }
-                        Some(established::Event::NotificationsOutReset { id, .. }) => {
-                            let outer_substream_id =
-                                outbound_substreams_reverse.remove(&id).unwrap();
-                            outbound_substreams_map.remove(&outer_substream_id);
-                            self.pending_messages.push_back(
-                                ConnectionToCoordinatorInner::NotificationsOutReset {
-                                    id: outer_substream_id,
-                                },
-                            );
-                        }
-                        Some(established::Event::PingOutSuccess) => {
-                            self.pending_messages
-                                .push_back(ConnectionToCoordinatorInner::PingOutSuccess);
-                        }
-                        Some(established::Event::PingOutFailed) => {
-                            self.pending_messages
-                                .push_back(ConnectionToCoordinatorInner::PingOutFailed);
-                        }
-                        None => {}
-                    }
-
-                    self.connection = SingleStreamConnectionTaskInner::Established {
-                        established: connection,
-                        outbound_substreams_map,
-                        outbound_substreams_reverse,
-                    };
-                }
-                Err(_err) => {
-                    // TODO: provide error
-                    self.pending_messages
-                        .push_back(ConnectionToCoordinatorInner::StartShutdown);
-                    self.pending_messages
-                        .push_back(ConnectionToCoordinatorInner::ShutdownFinished);
-                    self.connection = SingleStreamConnectionTaskInner::ShutdownWaitingAck {
-                        was_api_reset: false,
-                    };
-                }
-            },
-
-            SingleStreamConnectionTaskInner::Handshake {
-                mut handshake,
-                randomness_seed,
-                timeout,
-                noise_key,
-                notification_protocols,
-                request_response_protocols,
-                ping_protocol,
-            } => {
-                // Check that the handshake isn't taking too long.
-                //
-                // Note that we check this condition before looking into the incoming data,
-                // and it is possible for the buffers to contain the data that leads to the
-                // handshake being finished. If that is the case, however, it is impossible to
-                // know whether this data arrived before or after the timeout.
-                // Whether to put this check before or after reading the buffer is a choice
-                // between having false negatives or having false positives for the timeout.
-                // We are more strict than necessary by having the check before, but this
-                // guarantees that no horrendously slow connections can accidentally make their
-                // way through.
-                if timeout < read_write.now {
-                    // TODO: provide error: ConnectionError::Handshake(HandshakeError::Timeout)
-                    self.pending_messages
-                        .push_back(ConnectionToCoordinatorInner::StartShutdown);
-                    self.pending_messages
-                        .push_back(ConnectionToCoordinatorInner::ShutdownFinished);
-                    self.connection = SingleStreamConnectionTaskInner::ShutdownWaitingAck {
-                        was_api_reset: false,
-                    };
-                    return;
-                }
-
-                // `read_write()` should be called again as soon as possible after `timeout` in
-                // order for the check above to work.
-                read_write.wake_up_after(&timeout);
-
-                loop {
-                    let (read_before, written_before) =
-                        (read_write.read_bytes, read_write.written_bytes);
-
-                    let result = match handshake.read_write(read_write) {
-                        Ok(rw) => rw,
-                        Err(_err) => {
-                            // TODO: provide error: ConnectionError::Handshake(HandshakeError::Protocol(err))
-                            self.pending_messages
-                                .push_back(ConnectionToCoordinatorInner::StartShutdown);
-                            self.pending_messages
-                                .push_back(ConnectionToCoordinatorInner::ShutdownFinished);
-                            self.connection = SingleStreamConnectionTaskInner::ShutdownWaitingAck {
-                                was_api_reset: false,
-                            };
-                            return;
-                        }
-                    };
-
-                    match result {
-                        handshake::Handshake::Healthy(updated_handshake)
-                            if (read_before, written_before)
-                                == (read_write.read_bytes, read_write.written_bytes) =>
-                        {
-                            self.connection = SingleStreamConnectionTaskInner::Handshake {
-                                handshake: updated_handshake,
-                                randomness_seed,
-                                timeout,
-                                noise_key,
-                                notification_protocols,
-                                request_response_protocols,
-                                ping_protocol,
-                            };
-                            break;
-                        }
-                        handshake::Handshake::Healthy(updated_handshake) => {
-                            handshake = updated_handshake;
-                        }
-                        handshake::Handshake::Success {
-                            remote_peer_id,
-                            connection,
-                        } => {
-                            self.pending_messages.push_back(
-                                ConnectionToCoordinatorInner::HandshakeFinished(remote_peer_id),
-                            );
-
-                            self.connection = SingleStreamConnectionTaskInner::Established {
-                                established: connection.into_connection(established::Config {
-                                    notifications_protocols: notification_protocols
-                                        .iter()
-                                        .flat_map(|net| {
-                                            let max_handshake_size = net.config.max_handshake_size;
-                                            let max_notification_size =
-                                                net.config.max_notification_size;
-                                            iter::once(&net.config.protocol_name)
-                                                .chain(net.config.fallback_protocol_names.iter())
-                                                .map(move |name| {
-                                                    established::ConfigNotifications {
-                                                        name: name.clone(), // TODO: cloning :-/
-                                                        max_handshake_size,
-                                                        max_notification_size,
-                                                    }
-                                                })
-                                        })
-                                        .collect(),
-                                    request_protocols: request_response_protocols.to_vec(), // TODO: overhead
-                                    randomness_seed,
-                                    ping_protocol: ping_protocol.to_string(), // TODO: cloning :-/
-                                    ping_interval: Duration::from_secs(20),   // TODO: hardcoded
-                                    ping_timeout: Duration::from_secs(10),    // TODO: hardcoded
-                                    first_out_ping: read_write.now.clone() + Duration::from_secs(2), // TODO: hardcoded
-                                }),
-                                outbound_substreams_map:
-                                    hashbrown::HashMap::with_capacity_and_hasher(
-                                        0,
-                                        Default::default(),
-                                    ), // TODO: capacity?
-                                outbound_substreams_reverse:
-                                    hashbrown::HashMap::with_capacity_and_hasher(
-                                        0,
-                                        Default::default(),
-                                    ), // TODO: capacity?
-                            };
-                            break;
-                        }
-                        handshake::Handshake::NoiseKeyRequired(key) => {
-                            handshake = key.resume(&noise_key);
-                        }
-                    }
-                }
-            }
-
-            c @ (SingleStreamConnectionTaskInner::ShutdownWaitingAck {
-                was_api_reset: false,
-            }
-            | SingleStreamConnectionTaskInner::ShutdownAcked {
-                was_api_reset: false,
-            }) => {
-                // The user might legitimately call this function after the connection has
-                // already shut down. In that case, just do nothing.
-                self.connection = c;
-
-                // This might have been done already during the shutdown process, but we do it
-                // again just in case.
-                read_write.close_write();
-            }
-
-            SingleStreamConnectionTaskInner::ShutdownWaitingAck {
-                was_api_reset: true,
-            }
-            | SingleStreamConnectionTaskInner::ShutdownAcked {
-                was_api_reset: true,
-            } => {
-                // As documented, it is illegal to call this function after calling `reset()`.
-                panic!()
-            }
-
-            SingleStreamConnectionTaskInner::Poisoned => unreachable!(),
-        }
-    }
-}
-
-/// State machine dedicated to a single multi-stream connection.
-pub struct MultiStreamConnectionTask<TNow, TSubId> {
-    connection: MultiStreamConnectionTaskInner<TNow, TSubId>,
-}
-enum MultiStreamConnectionTaskInner<TNow, TSubId> {
-    /// Connection has been fully established.
-    Established {
-        // TODO: user data of request redundant with the substreams mapping below
-        established: established::MultiStream<TNow, TSubId, SubstreamId, ()>,
-
-        /// Because outgoing substream ids are assigned by the coordinator, we maintain a mapping
-        /// of the "outer ids" to "inner ids".
-        outbound_substreams_map:
-            hashbrown::HashMap<SubstreamId, established::SubstreamId, fnv::FnvBuildHasher>,
-
-        /// Reverse mapping.
-        // TODO: could be user datas in established?
-        outbound_substreams_reverse:
-            hashbrown::HashMap<established::SubstreamId, SubstreamId, fnv::FnvBuildHasher>,
-    },
-
-    /// Connection has finished its shutdown. A [`ConnectionToCoordinatorInner::ShutdownFinished`]
-    /// message has been sent and is waiting to be acknowledged.
-    ShutdownWaitingAck {
-        /// If true, [`MultiStreamConnectionTask::reset`] has been called. This doesn't modify any
-        /// of the behavior but is used to make sure that the API is used correctly.
-        was_api_reset: bool,
-
-        /// `true` if the [`ConnectionToCoordinatorInner::StartShutdown`] message has already
-        /// been sent to the coordinator.
-        start_shutdown_message_sent: bool,
-
-        /// `true` if the [`ConnectionToCoordinatorInner::ShutdownFinished`] message has already
-        /// been sent to the coordinator.
-        shutdown_finish_message_sent: bool,
-    },
-
-    /// Connection has finished its shutdown and its shutdown has been acknowledged. There is
-    /// nothing more to do except stop the connection task.
-    ShutdownAcked {
-        /// If true, [`MultiStreamConnectionTask::reset`] has been called. This doesn't modify any
-        /// of the behavior but is used to make sure that the API is used correctly.
-        was_api_reset: bool,
-    },
-}
-
-impl<TNow, TSubId> MultiStreamConnectionTask<TNow, TSubId>
-where
-    TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,
-    TSubId: Clone + PartialEq + Eq + Hash,
-{
-    /// Pulls a message to send back to the coordinator.
-    ///
-    /// This function takes ownership of `self` and optionally yields it back. If the first
-    /// option contains `None`, then no more message will be generated and the
-    /// [`MultiStreamConnectionTask`] has vanished. This will happen after the connection has been
-    /// shut down or reset.
-    /// It is possible for `self` to not be yielded back even if substreams are still open, in
-    /// which case the API user should abruptly reset the connection, for example by sending a
-    /// TCP RST flag.
-    ///
-    /// If any message is returned, it is the responsibility of the API user to send it to the
-    /// coordinator by calling [`Network::inject_connection_message`].
-    /// Do not attempt to buffer the message being returned, as it would work against the
-    /// back-pressure strategy used internally. As soon as a message is returned, it should be
-    /// delivered. If the coordinator is busy at the moment a message should be delivered, then
-    /// the entire thread of execution dedicated to this [`MultiStreamConnectionTask`] should be
-    /// paused until the coordinator is ready and the message delivered.
-    ///
-    /// Messages aren't generated spontaneously. In other words, you don't need to periodically
-    /// call this function just in case there's a new message. Messages are always generated after
-    /// [`MultiStreamConnectionTask::substream_read_write`],
-    /// [`MultiStreamConnectionTask::add_substream`], or [`MultiStreamConnectionTask::reset`]
-    /// has been called. Multiple messages can happen in a row.
-    ///
-    /// Because this function frees space in a buffer, calling
-    /// [`MultiStreamConnectionTask::ready_substreams`] and processing substreams again after it
-    /// has returned might read/write more data and generate an event again. In other words,
-    /// the API user should call
-    ///  [`MultiStreamConnectionTask::ready_substreams`] and
-    /// [`MultiStreamConnectionTask::substream_read_write`], and
-    /// [`MultiStreamConnectionTask::pull_message_to_coordinator`] repeatedly in a loop until no
-    /// more message is generated.
-    pub fn pull_message_to_coordinator(
-        mut self,
-    ) -> (Option<Self>, Option<ConnectionToCoordinator>) {
-        match &mut self.connection {
-            MultiStreamConnectionTaskInner::Established {
-                established,
-                outbound_substreams_map,
-                outbound_substreams_reverse,
-            } => {
-                let event = match established.pull_event() {
-                    Some(established::Event::InboundError(err)) => {
-                        Some(ConnectionToCoordinatorInner::InboundError(err))
-                    }
-                    Some(established::Event::RequestIn {
-                        id,
-                        protocol_index,
-                        request,
-                    }) => Some(ConnectionToCoordinatorInner::RequestIn {
-                        id,
-                        protocol_index,
-                        request,
-                    }),
-                    Some(established::Event::Response { id, response, .. }) => {
-                        let outer_substream_id = outbound_substreams_reverse.remove(&id).unwrap();
-                        outbound_substreams_map.remove(&outer_substream_id).unwrap();
-                        Some(ConnectionToCoordinatorInner::Response {
-                            response,
-                            id: outer_substream_id,
-                        })
-                    }
-                    Some(established::Event::NotificationsInOpen {
-                        id,
-                        protocol_index,
-                        handshake,
-                    }) => Some(ConnectionToCoordinatorInner::NotificationsInOpen {
-                        id,
-                        protocol_index,
-                        handshake,
-                    }),
-                    Some(established::Event::NotificationsInOpenCancel { id, .. }) => {
-                        Some(ConnectionToCoordinatorInner::NotificationsInOpenCancel { id })
-                    }
-                    Some(established::Event::NotificationIn { id, notification }) => {
-                        Some(ConnectionToCoordinatorInner::NotificationIn { id, notification })
-                    }
-                    Some(established::Event::NotificationsInClose { id, outcome, .. }) => {
-                        Some(ConnectionToCoordinatorInner::NotificationsInClose { id, outcome })
-                    }
-                    Some(established::Event::NotificationsOutResult { id, result }) => {
-                        let outer_substream_id = *outbound_substreams_reverse.get(&id).unwrap();
-
-                        if result.is_err() {
-                            outbound_substreams_map.remove(&outer_substream_id);
-                            outbound_substreams_reverse.remove(&id);
-                        }
-
-                        Some(ConnectionToCoordinatorInner::NotificationsOutResult {
-                            id: outer_substream_id,
-                            result: result.map_err(|(err, _)| NotificationsOutErr::Substream(err)),
-                        })
-                    }
-                    Some(established::Event::NotificationsOutCloseDemanded { id }) => {
-                        let outer_substream_id = *outbound_substreams_reverse.get(&id).unwrap();
-                        Some(
-                            ConnectionToCoordinatorInner::NotificationsOutCloseDemanded {
-                                id: outer_substream_id,
-                            },
-                        )
-                    }
-                    Some(established::Event::NotificationsOutReset { id, .. }) => {
-                        let outer_substream_id = outbound_substreams_reverse.remove(&id).unwrap();
-                        outbound_substreams_map.remove(&outer_substream_id);
-                        Some(ConnectionToCoordinatorInner::NotificationsOutReset {
-                            id: outer_substream_id,
-                        })
-                    }
-                    Some(established::Event::PingOutSuccess) => {
-                        Some(ConnectionToCoordinatorInner::PingOutSuccess)
-                    }
-                    Some(established::Event::PingOutFailed) => {
-                        Some(ConnectionToCoordinatorInner::PingOutFailed)
-                    }
-                    None => None,
-                };
-
-                (
-                    Some(self),
-                    event.map(|ev| ConnectionToCoordinator { inner: ev }),
-                )
-            }
-            MultiStreamConnectionTaskInner::ShutdownWaitingAck {
-                start_shutdown_message_sent,
-                shutdown_finish_message_sent,
-                ..
-            } => {
-                if !*start_shutdown_message_sent {
-                    debug_assert!(!*shutdown_finish_message_sent);
-                    *start_shutdown_message_sent = true;
-                    (
-                        Some(self),
-                        Some(ConnectionToCoordinator {
-                            inner: ConnectionToCoordinatorInner::StartShutdown,
-                        }),
-                    )
-                } else if !*shutdown_finish_message_sent {
-                    debug_assert!(*start_shutdown_message_sent);
-                    *shutdown_finish_message_sent = true;
-                    (
-                        Some(self),
-                        Some(ConnectionToCoordinator {
-                            inner: ConnectionToCoordinatorInner::ShutdownFinished,
-                        }),
-                    )
-                } else {
-                    (Some(self), None)
-                }
-            }
-            MultiStreamConnectionTaskInner::ShutdownAcked { .. } => (None, None),
-        }
-    }
-
-    /// Injects a message that has been pulled using [`Network::pull_message_to_connection`].
-    ///
-    /// Calling this function might generate data to send to the connection. You should call
-    /// [`MultiStreamConnectionTask::desired_outbound_substreams`] and
-    /// [`MultiStreamConnectionTask::ready_substreams`] after this function has returned.
-    pub fn inject_coordinator_message(&mut self, message: CoordinatorToConnection<TNow>) {
-        match (message.inner, &mut self.connection) {
-            (
-                CoordinatorToConnectionInner::StartRequest {
-                    request_data,
-                    timeout,
-                    protocol_index,
-                    substream_id,
-                },
-                MultiStreamConnectionTaskInner::Established {
-                    established,
-                    outbound_substreams_map,
-                    outbound_substreams_reverse,
-                },
-            ) => {
-                let inner_substream_id =
-                    established.add_request(protocol_index, request_data, timeout, substream_id);
-                let _prev_value = outbound_substreams_map.insert(substream_id, inner_substream_id);
-                debug_assert!(_prev_value.is_none());
-                let _prev_value =
-                    outbound_substreams_reverse.insert(inner_substream_id, substream_id);
-                debug_assert!(_prev_value.is_none());
-            }
-            (
-                CoordinatorToConnectionInner::OpenOutNotifications {
-                    handshake,
-                    now,
-                    overlay_network_index,
-                    substream_id: outer_substream_id,
-                },
-                MultiStreamConnectionTaskInner::Established {
-                    established,
-                    outbound_substreams_map,
-                    outbound_substreams_reverse,
-                },
-            ) => {
-                let inner_substream_id = established.open_notifications_substream(
-                    now,
-                    overlay_network_index,
-                    handshake,
-                    (),
-                );
-
-                let _prev_value =
-                    outbound_substreams_map.insert(outer_substream_id, inner_substream_id);
-                debug_assert!(_prev_value.is_none());
-                let _prev_value =
-                    outbound_substreams_reverse.insert(inner_substream_id, outer_substream_id);
-                debug_assert!(_prev_value.is_none());
-            }
-            (
-                CoordinatorToConnectionInner::CloseOutNotifications { substream_id },
-                MultiStreamConnectionTaskInner::Established {
-                    established,
-                    outbound_substreams_map,
-                    outbound_substreams_reverse,
-                },
-            ) => {
-                // It is possible that the remote has closed the outbound notification substream
-                // while the `CloseOutNotifications` message was being delivered, or that the API
-                // user close the substream before the message about the substream being closed
-                // was delivered to the coordinator.
-                if let Some(inner_substream_id) = outbound_substreams_map.remove(&substream_id) {
-                    outbound_substreams_reverse.remove(&inner_substream_id);
-                    established.close_notifications_substream(inner_substream_id);
-                }
-            }
-            (
-                CoordinatorToConnectionInner::QueueNotification {
-                    substream_id,
-                    notification,
-                },
-                MultiStreamConnectionTaskInner::Established {
-                    established,
-                    outbound_substreams_map,
-                    ..
-                },
-            ) => {
-                // It is possible that the remote has closed the outbound notification substream
-                // while a `QueueNotification` message was being delivered, or that the API user
-                // queued a notification before the message about the substream being closed was
-                // delivered to the coordinator.
-                // If that happens, we intentionally silently discard the message, causing the
-                // notification to not be sent. This is consistent with the guarantees about
-                // notifications delivered that are documented in the public API.
-                if let Some(inner_substream_id) = outbound_substreams_map.get(&substream_id) {
-                    established.write_notification_unbounded(*inner_substream_id, notification);
-                }
-            }
-            (
-                CoordinatorToConnectionInner::AnswerRequest {
-                    substream_id,
-                    response,
-                },
-                MultiStreamConnectionTaskInner::Established { established, .. },
-            ) => match established.respond_in_request(substream_id, response) {
-                Ok(()) => {}
-                Err(established::RespondInRequestError::SubstreamClosed) => {
-                    // As documented, answering an obsolete request is simply ignored.
-                }
-            },
-            (
-                CoordinatorToConnectionInner::AcceptInNotifications {
-                    substream_id,
-                    handshake,
-                },
-                MultiStreamConnectionTaskInner::Established { established, .. },
-            ) => {
-                // TODO: must verify that the substream is still valid
-                established.accept_in_notifications_substream(substream_id, handshake, ());
-            }
-            (
-                CoordinatorToConnectionInner::RejectInNotifications { substream_id },
-                MultiStreamConnectionTaskInner::Established { established, .. },
-            ) => {
-                // TODO: must verify that the substream is still valid
-                established.reject_in_notifications_substream(substream_id);
-            }
-            (
-                CoordinatorToConnectionInner::StartShutdown { .. },
-                MultiStreamConnectionTaskInner::Established { .. },
-            ) => {
-                // TODO: implement proper shutdown
-                self.connection = MultiStreamConnectionTaskInner::ShutdownWaitingAck {
-                    start_shutdown_message_sent: false,
-                    shutdown_finish_message_sent: false,
-                    was_api_reset: false,
-                };
-            }
-            (
-                CoordinatorToConnectionInner::AcceptInNotifications { .. }
-                | CoordinatorToConnectionInner::RejectInNotifications { .. }
-                | CoordinatorToConnectionInner::AnswerRequest { .. }
-                | CoordinatorToConnectionInner::OpenOutNotifications { .. }
-                | CoordinatorToConnectionInner::CloseOutNotifications { .. }
-                | CoordinatorToConnectionInner::QueueNotification { .. },
-                MultiStreamConnectionTaskInner::ShutdownAcked { .. },
-            ) => unreachable!(),
-            (
-                CoordinatorToConnectionInner::AcceptInNotifications { .. }
-                | CoordinatorToConnectionInner::RejectInNotifications { .. }
-                | CoordinatorToConnectionInner::AnswerRequest { .. }
-                | CoordinatorToConnectionInner::OpenOutNotifications { .. }
-                | CoordinatorToConnectionInner::CloseOutNotifications { .. }
-                | CoordinatorToConnectionInner::QueueNotification { .. },
-                MultiStreamConnectionTaskInner::ShutdownWaitingAck { .. },
-            ) => {
-                // There might still be some messages coming from the coordinator after the
-                // connection task has sent a message indicating that it has shut down. This is
-                // due to the concurrent nature of the API and doesn't indicate a bug. These
-                // messages are simply ignored by the connection task.
-            }
-            (
-                CoordinatorToConnectionInner::ShutdownFinishedAck,
-                MultiStreamConnectionTaskInner::ShutdownWaitingAck {
-                    start_shutdown_message_sent,
-                    shutdown_finish_message_sent,
-                    was_api_reset: was_reset,
-                },
-            ) => {
-                debug_assert!(*start_shutdown_message_sent && *shutdown_finish_message_sent);
-                self.connection = MultiStreamConnectionTaskInner::ShutdownAcked {
-                    was_api_reset: *was_reset,
-                };
-            }
-            _ => todo!(), // TODO:
-        }
-    }
-
-    /// Returns the number of new outbound substreams that the state machine would like to see
-    /// opened.
-    ///
-    /// This value doesn't change automatically over time but only after a call to
-    /// [`MultiStreamConnectionTask::substream_read_write`],
-    /// [`MultiStreamConnectionTask::inject_coordinator_message`],
-    /// [`MultiStreamConnectionTask::add_substream`], or
-    /// [`MultiStreamConnectionTask::reset_substream`].
-    ///
-    /// Note that the user is expected to track the number of substreams that are currently being
-    /// opened. For example, if this function returns 2 and there are already 2 substreams
-    /// currently being opened, then there is no need to open any additional one.
-    pub fn desired_outbound_substreams(&self) -> u32 {
-        match &self.connection {
-            MultiStreamConnectionTaskInner::Established { established, .. } => {
-                established.desired_outbound_substreams()
-            }
-            _ => 0,
-        }
-    }
-
-    /// Notifies the state machine that a new substream has been opened.
-    ///
-    /// `inbound` indicates whether the substream has been opened by the remote (`true`) or
-    /// locally (`false`).
-    ///
-    /// If `inbound` is `false`, then the value returned by
-    /// [`MultiStreamConnectionTask::desired_outbound_substreams`] will decrease by one.
-    ///
-    /// # Panic
-    ///
-    /// Panics if there already exists a substream with an identical identifier.
-    ///
-    pub fn add_substream(&mut self, id: TSubId, inbound: bool) {
-        match &mut self.connection {
-            MultiStreamConnectionTaskInner::Established { established, .. } => {
-                established.add_substream(id, inbound)
-            }
-            _ => {
-                // TODO: reset the substream or something?
-            }
-        }
-    }
-
-    /// Returns a list of substreams that the state machine would like to see reset. The user is
-    /// encouraged to call [`MultiStreamConnectionTask::substream_read_write`] with this list of
-    /// substream.
-    ///
-    /// This value doesn't change automatically over time but only after a call to
-    /// [`MultiStreamConnectionTask::substream_read_write`],
-    /// [`MultiStreamConnectionTask::inject_coordinator_message`],
-    /// [`MultiStreamConnectionTask::add_substream`], or
-    /// [`MultiStreamConnectionTask::reset_substream`].
-    ///
-    /// > **Note**: An example situation is: a notification is queued, which leads to a message
-    /// >           being sent to a connection task, which, once injected, leads to a notifications
-    /// >           substream being "ready" because it needs to send more data.
-    pub fn ready_substreams(&self) -> impl Iterator<Item = &TSubId> {
-        match &self.connection {
-            MultiStreamConnectionTaskInner::Established { established, .. } => {
-                either::Left(established.ready_substreams())
-            }
-            _ => either::Right(iter::empty()),
-        }
-    }
-
-    /// Sets the state of the connection to "reset".
-    ///
-    /// This should be called if the remote abruptly closes the connection, such as with a TCP/IP
-    /// RST flag.
-    ///
-    /// After this function has been called, it is illegal to call
-    /// [`MultiStreamConnectionTask::substream_read_write`] or
-    /// [`MultiStreamConnectionTask::reset`] again.
-    ///
-    /// Calling this function might have generated messages for the coordinator.
-    /// [`MultiStreamConnectionTask::pull_message_to_coordinator`] should be called afterwards in
-    /// order to process these messages.
-    ///
-    /// # Panic
-    ///
-    /// Panics if [`MultiStreamConnectionTask::reset`] has been called in the past.
-    ///
-    pub fn reset(&mut self) {
-        // It is illegal to call `reset` a second time. Verify that the user didn't do this.
-        if let MultiStreamConnectionTaskInner::ShutdownWaitingAck {
-            was_api_reset: true,
-            ..
-        }
-        | MultiStreamConnectionTaskInner::ShutdownAcked {
-            was_api_reset: true,
-        } = self.connection
-        {
-            panic!()
-        }
-
-        self.connection = MultiStreamConnectionTaskInner::ShutdownWaitingAck {
-            was_api_reset: true,
-            shutdown_finish_message_sent: false,
-            start_shutdown_message_sent: false,
-        };
-    }
-
-    /// Immediately destroys the substream with the given identifier.
-    ///
-    /// The given identifier is now considered invalid by the state machine.
-    ///
-    /// # Panic
-    ///
-    /// Panics if there is no substream with that identifier.
-    ///
-    pub fn reset_substream(&mut self, substream_id: &TSubId) {
-        match &mut self.connection {
-            MultiStreamConnectionTaskInner::Established { established, .. } => {
-                established.reset_substream(substream_id)
-            }
-            _ => {
-                // TODO: panic if substream id invalid?
-            }
-        }
-    }
-
-    /// Reads/writes data on the substream.
-    ///
-    /// If the method returns `true`, then the substream is now considered dead according to the
-    /// state machine and its identifier is now invalid. If the reading or writing side of the
-    /// substream was still open, then the user should reset that substream.
-    ///
-    /// # Panic
-    ///
-    /// Panics if there is no substream with that identifier.
-    ///
-    // TODO: better return value
-    pub fn substream_read_write(
-        &mut self,
-        substream_id: &TSubId,
-        read_write: &'_ mut ReadWrite<'_, TNow>,
-    ) -> bool {
-        match &mut self.connection {
-            MultiStreamConnectionTaskInner::Established { established, .. } => {
-                established.substream_read_write(substream_id, read_write)
-            }
-            _ => {
-                // TODO: panic if substream id invalid?
-                true
-            }
-        }
-    }
 }
 
 pub struct ConnectionToCoordinator {
